requirements.txt
pandas
sqlalchemy
psycopg2
pyarrow
python-dotenv